lispの関数を本体から分離
