+++ scheme/util.go
@@ -1,0 +1,51 @@
+/*
+   Go lang 3rd study program.
+   This is prototype program mini scheme subset.
+
+   hidekuno@gmail.com
+*/
+package scheme
+
+import (
+	"fmt"
+	"os"
+	"reflect"
+	"strconv"
+	"time"
+)
+
+// Build Global environement.
+func buildUtilFunc() {
+
+	buildInFuncTbl["identity"] = func(exp []Expression, env *SimpleEnv) (Expression, error) {
+		if len(exp) != 1 {
+			return nil, NewRuntimeError("E1007", strconv.Itoa(len(exp)))
+		}
+		return eval(exp[0], env)
+	}
+	buildInFuncTbl["time"] = func(exp []Expression, env *SimpleEnv) (Expression, error) {
+		if len(exp) != 1 {
+			return nil, NewRuntimeError("E1007", strconv.Itoa(len(exp)))
+		}
+		t0 := time.Now()
+		e, err := eval(exp[0], env)
+		t1 := time.Now()
+		fmt.Println(t1.Sub(t0))
+		return e, err
+	}
+	//srfi-98
+	buildInFuncTbl["get-environment-variable"] = func(exp []Expression, env *SimpleEnv) (Expression, error) {
+		if len(exp) != 1 {
+			return nil, NewRuntimeError("E1007", strconv.Itoa(len(exp)))
+		}
+		v, err := eval(exp[0], env)
+		if err != nil {
+			return v, err
+		}
+		s, ok := v.(*String)
+		if !ok {
+			return nil, NewRuntimeError("E1015", reflect.TypeOf(exp[0]).String())
+		}
+		return NewString(os.Getenv(s.Value)), nil
+	}
+}
